move Calendar and Modal into utility components
