+++ client/src/components/habits/Habits/Habits.tsx
@@ -1,7 +1,7 @@
 import DetailedHabit from "@/components/habits/DetailedHabit/DetailedHabit";
 import Habit from "@/components/habits/Habits/Habit";
 import useDetailedHabitModal from "@/components/habits/Habits/useDetailedHabitModal";
-import Modal from "@/components/Modal/Modal";
+import Modal from "@/components/utility/Modal/Modal";
 import L from "@/lib/theme/components/List.style";
 import type { HabitWithPossiblySyntheticEntries } from "@/types/server/habit.types";
 import type { ById } from "@/types/server/utility.types";
