+++ client/src/components/utility/Calendar/build-calendar-rows.ts
@@ -1,5 +1,5 @@
-import type { Row, Rows, WeekStartDay } from "@/components/Calendar/calendar.types";
 import type { Dayjs } from "dayjs";
+import type { Row, Rows, WeekStartDay } from "./calendar.types";
 
 /**
  * Every row in a calendar has 7 cells (1 cell for each weekday), but the first
